requirement.txt
openai
langgraph
streamlit
PyPDF2
google-genai
google-generativeai
